enable fpu in the whole softirq context
Now fpu can be enabled in the whole softirq context. Unlike the old
version, the new kernel raises softirq even in the boot phase, when fpu
related stuff (such as registers or process fpu state) is not ready for
manipulation (otherwise exception raised), so it should be enabled in
the first kernel_fpu_begin() (i.e. the first possible crypto API call
happens in the softirq context) instead of __do_softirq().